Bug fix, 0.2.1 release.
Dependencies were being incorrectly validated many times, even if they
were already in the validated fields list. This would cause None to be
set as a dependency's value if it had already been validated.

This commit fixes this problem, and ensures a valid test case exists to
fix it for good.